sign-language-deployment/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.1.0
boto3==1.34.0
pillow==9.5.0
python-multipart==0.0.6
numpy==1.24.4
opencv-python==4.7.0.72

